重构：提取推进器控制系统为单独类
